added authf

+++ src/main/java/com/stepping/step5/service/CurrentUserService.java
@@ -6,5 +6,5 @@ import com.stepping.step5.models.CurrentUser;
  * Created by re5 on 24.10.16.
  */
 public interface CurrentUserService {
-    boolean canAccessUser(CurrentUser currentUser, Integer userId);
+    boolean canAccessUser(CurrentUser currentUser, Long userId);
 }
